# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type BookTestament = typeof BookTestament[keyof typeof BookTestament];


export const BookTestament = {
  old: 'old',
  new: 'new',
} as const;

export interface Book {
  id: string;
  slug: string;
  name: string;
  testament: BookTestament;
  category: string;
  chapterCount: number;
}

export interface Verse {
  id: string;
  bookSlug: string;
  bookName: string;
  chapter: number;
  verse: number;
  text: string;
  reference: string;
  translation: string;
}

export interface Chapter {
  book: Book;
  chapter: number;
  verses: Verse[];
}

export interface PersonSummary {
  slug: string;
  name: string;
  role: string;
  description: string;
}

export type PlaceSummaryConfidence = typeof PlaceSummaryConfidence[keyof typeof PlaceSummaryConfidence];


export const PlaceSummaryConfidence = {
  known: 'known',
  traditional: 'traditional',
  approximate: 'approximate',
  disputed: 'disputed',
} as const;

export interface PlaceSummary {
  slug: string;
  name: string;
  region: string;
  confidence: PlaceSummaryConfidence;
}

export interface EventSummary {
  slug: string;
  name: string;
  summary: string;
  period: string;
}

export interface TimelineEvent {
  slug: string;
  title: string;
  period: string;
  dateLabel: string;
  description: string;
  scriptureReferences: string[];
}

export type Person = PersonSummary & {
  scriptureReferences: string[];
  places: PlaceSummary[];
  events: EventSummary[];
  timeline: TimelineEvent[];
};

export type Place = PlaceSummary & {
  ancientName: string;
  latitude: number;
  longitude: number;
  historicalNotes: string;
  scriptureReferences: string[];
  people: PersonSummary[];
  events: EventSummary[];
};

export type Event = EventSummary & {
  scriptureReferences: string[];
  people: PersonSummary[];
  places: PlaceSummary[];
  before: string;
  after: string;
};

export interface JourneySummary {
  slug: string;
  title: string;
  subtitle: string;
  stopCount: number;
  durationLabel: string;
}

export interface JourneyStop {
  title: string;
  location: string;
  description: string;
  scriptureReferences: string[];
  placeSlug: string;
}

export type Journey = JourneySummary & {
  stops: JourneyStop[];
};

export interface VerseExploration {
  verse: Verse;
  people: PersonSummary[];
  places: PlaceSummary[];
  events: EventSummary[];
  timeline: TimelineEvent[];
  relatedScripture: string[];
  context: string;
  beforeEvent: string;
  afterEvent: string;
}

export interface SearchResults {
  scripture: Verse[];
  people: PersonSummary[];
  places: PlaceSummary[];
  events: EventSummary[];
  journeys: JourneySummary[];
}

export interface PassageRequest {
  verseId: string;
  passage: string;
  reference: string;
}

export interface PassageQuestion {
  verseId: string;
  passage: string;
  reference: string;
  question: string;
}

export interface SourceReference {
  label: string;
  reference: string;
}

export type AIExplanationLabel = typeof AIExplanationLabel[keyof typeof AIExplanationLabel];


export const AIExplanationLabel = {
  'AI-assisted_explanation': 'AI-assisted explanation',
} as const;

export interface AIExplanation {
  reference: string;
  scripture: string;
  simpleWords: string;
  whyItMatters: string;
  sources: SourceReference[];
  label: AIExplanationLabel;
}

export type AIAnswerLabel = typeof AIAnswerLabel[keyof typeof AIAnswerLabel];


export const AIAnswerLabel = {
  'AI-assisted_answer': 'AI-assisted answer',
} as const;

export interface AIAnswer {
  shortAnswer: string;
  scripture: string[];
  context: string;
  explanation: string;
  exploreNext: string[];
  sources: SourceReference[];
  label: AIAnswerLabel;
}

export interface ReadingProgress {
  bookSlug: string;
  chapter: number;
  verse: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percent: number;
  lastReference: string;
}

export interface ProgressInput {
  bookSlug: string;
  chapter: number;
  verse: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percent: number;
  lastReference: string;
}

export type SearchContentParams = {
/**
 * @minLength 2
 */
q: string;
};