# lib/api-zod/src/generated/types/aIAnswer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */
import type { AIAnswerLabel } from './aIAnswerLabel';
import type { SourceReference } from './sourceReference';

export interface AIAnswer {
  shortAnswer: string;
  scripture: string[];
  context: string;
  explanation: string;
  exploreNext: string[];
  sources: SourceReference[];
  label: AIAnswerLabel;
}