# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from './book';
import type { Verse } from './verse';

export interface Chapter {
  book: Book;
  chapter: number;
  verses: Verse[];
}