# lib/api-zod/src/generated/types/passageQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */

export interface PassageQuestion {
  verseId: string;
  passage: string;
  reference: string;
  question: string;
}