# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */
import type { EventSummary } from './eventSummary';
import type { PersonSummary } from './personSummary';
import type { PlaceSummary } from './placeSummary';
import type { TimelineEvent } from './timelineEvent';

export type Person = PersonSummary & {
  scriptureReferences: string[];
  places: PlaceSummary[];
  events: EventSummary[];
  timeline: TimelineEvent[];
};