# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */
import type { EventSummary } from './eventSummary';
import type { PersonSummary } from './personSummary';
import type { PlaceSummary } from './placeSummary';

export type Place = PlaceSummary & {
  ancientName: string;
  latitude: number;
  longitude: number;
  historicalNotes: string;
  scriptureReferences: string[];
  people: PersonSummary[];
  events: EventSummary[];
};