# lib/api-zod/src/generated/types/progressInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressInput {
  bookSlug: string;
  chapter: number;
  verse: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percent: number;
  lastReference: string;
}