# lib/api-zod/src/generated/types/sourceReference.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Christ: The Scroll, an interactive Bible exploration experience.
 * OpenAPI spec version: 0.1.0
 */

export interface SourceReference {
  label: string;
  reference: string;
}